+++ jdaviz/configs/cubeviz/plugins/sonify_data/sonify_data.vue
@@ -9,12 +9,23 @@
     :scroll_to.sync="scroll_to"
     :disabled_msg="disabled_msg">
 
-    <j-plugin-section-header>Sonify Cube</j-plugin-section-header>
+    <j-plugin-section-header>Cube Pre-Sonification Options</j-plugin-section-header>
     <v-alert v-if="!has_strauss" type="warning" style="margin-left: -12px; margin-right: -12px">
       To use Sonify Data, install strauss and restart Jdaviz. You can do this by running pip install strauss
       in the command line and then launching Jdaviz.
     </v-alert>
-
+    <v-row>
+      <j-docs-link>Choose the input cube, spectral subset and any advanced sonification options.</j-docs-link>
+    </v-row>
+    <plugin-dataset-select
+      :items="dataset_items"
+      :selected.sync="dataset_selected"
+      :show_if_single_entry="false"
+      label="Data"
+      api_hint="plg.dataset ="
+      :api_hints_enabled="api_hints_enabled"
+      hint="Select the data set."
+    />
     <plugin-subset-select
       :items="spectral_subset_items"
       :selected.sync="spectral_subset_selected"
@@ -24,7 +35,6 @@
       :api_hints_enabled="api_hints_enabled"
       hint="Select spectral region that defines the wavelength range."
     />
-
     <v-row>
       <v-expansion-panels accordion>
         <v-expansion-panel>
@@ -32,26 +42,6 @@
             <span style="padding: 6px">Advanced Sound Options</span>
           </v-expansion-panel-header>
           <v-expansion-panel-content class="plugin-expansion-panel-content">
-            <v-row>
-              <v-text-field
-                ref="sample_rate"
-                type="number"
-                label="Sample Rate"
-                v-model.number="sample_rate"
-                hint="The desired sample rate."
-                persistent-hint
-              ></v-text-field>
-            </v-row>
-            <v-row>
-              <v-text-field
-                ref="buffer_size"
-                type="number"
-                label="Buffer Size"
-                v-model.number="buffer_size"
-                hint="The desired buffer size."
-                persistent-hint
-              ></v-text-field>
-            </v-row>
             <v-row>
               <v-text-field
                 ref="audfrqmin"
@@ -92,13 +82,21 @@
                 persistent-hint
               ></v-text-field>
             </v-row>
-            <v-row>
+	    <v-row>
+              <v-switch
+                v-model="use_pccut"
+                label="Use Flux Percentile Cut?"
+                hint="Whether to only sonify flux above a min. percentile (else use absolute values)"
+                persistent-hint
+               ></v-switch>
+	    </v-row>
+            <v-row v-if="use_pccut">
               <v-text-field
                 ref="pccut"
                 type="number"
-                label="Flux Percentile Cut"
+                label="Flux Percentile Cut Value"
                 v-model.number="pccut"
-                hint="The minimum flux percentile to be heard."
+                hint="The minimum percentile to be heard."
                 persistent-hint
               ></v-text-field>
             </v-row>
@@ -138,6 +136,7 @@
         Refresh Device List
       </plugin-action-button>
     </v-row>
+    <j-plugin-section-header>Live Sound Options</j-plugin-section-header>
     <v-row>
       <v-select
         :menu-props="{ left: true }"
@@ -145,7 +144,7 @@
         :items="sound_devices_items"
         v-model="sound_devices_selected"
         label="Sound device"
-        hint="Device which sound will be output from. Must be selected BEFORE cube is sonified."
+        hint="Device which sound will be output from."
         persistent-hint
         ></v-select>
     </v-row>
@@ -153,5 +152,6 @@
         Volume
         <glue-throttled-slider label="Volume" wait="300" max="100" step="1" :value.sync="volume" hide-details class="no-hint" />
     </v-row>
+   </v-row>
  </j-tray-plugin>
 </template>